requirements.txt
Python==3.9.5
Django==3.2.8
djangorestframework==3.12.4
